[feat]: se agrega la funcionalidad de consultar libros por prestamo

+++ src/main/java/com/dancmo/project/libraryud/persistence/repository/AutorRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface AutorRepository extends JpaRepository<Autor, Long> {
+public interface AutorRepository extends JpaRepository<Autor, Integer> {
     List<Autor> findAllByNombre(String autor);
 
 }
